+++ resources/views/components/list/index.blade.php
@@ -1,0 +1,17 @@
+@props([
+    /** @var string $container */
+    'container' => 'ul',
+
+    /** @var bool $flush */
+    'flush' => false,
+
+    /** @var bool $horizontal */
+    'horizontal' => false,
+])
+<{{ $container }} {{ $attributes
+    ->class([
+        $horizontal ? 'list-group list-group-horizontal' : 'list-group',
+        'list-group-flush' => $flush,
+    ]) }}>
+    {{ $slot }}
+</{{ $container }}>