+++ tests/Feature/Button/ButtonTest.php
@@ -3,11 +3,14 @@
 namespace Portavice\Bladestrap\Tests\Feature\Button;
 
 use Portavice\Bladestrap\Tests\Feature\ComponentTestCase;
+use Portavice\Bladestrap\Tests\Traits\TestsVariants;
 
 class ButtonTest extends ComponentTestCase
 {
+    use TestsVariants;
+
     /**
-     * @dataProvider dataProviderForVariants
+     * @dataProvider variants
      */
     public function testButtonsRenderCorrectly(string $buttonClass, ?string $variant): void
     {
@@ -16,14 +19,14 @@ class ButtonTest extends ComponentTestCase
             $this->bladeView(
                 sprintf(
                     '<x-bs::button %s>Button title</x-bs::button>',
-                    isset($variant) ? sprintf('variant="%s"', $variant) : ''
+                    self::makeVariantAttribute($variant)
                 )
             )
         );
     }
 
     /**
-     * @dataProvider dataProviderForVariants
+     * @dataProvider variants
      */
     public function testDisabledButtonsRenderCorrectly(string $buttonClass, ?string $variant): void
     {
@@ -31,20 +34,18 @@ class ButtonTest extends ComponentTestCase
             '<button aria-disabled="true" tabindex="-1" class="btn ' . $buttonClass . ' disabled" disabled>Button title</button>',
             $this->bladeView(
                 sprintf(
-                    '<x-bs::button %s disabled="true">Button title</x-bs::button>',
-                    isset($variant) ? sprintf('variant="%s"', $variant) : ''
+                    '<x-bs::button %s :disabled="true">Button title</x-bs::button>',
+                    self::makeVariantAttribute($variant)
                 )
             )
         );
     }
 
-    public static function dataProviderForVariants(): array
+    public static function variants(): array
     {
         return [
             ['btn-primary', null],
-            ['btn-primary', 'primary'],
-            ['btn-secondary', 'secondary'],
-            ['btn-danger', 'danger'],
+            ...self::makeDataProvider('btn-'),
             ['btn-link', 'link'],
         ];
     }